+++ client/cmdhw.c
@@ -498,11 +498,14 @@ void pm3_version(bool verbose) {
         PrintAndLogEx(NORMAL, "\n" _BLUE_(" [ Proxmark3 RFID instrument ]") "\n");
         PrintAndLogEx(NORMAL, "\n [ CLIENT ]");
         PrintAndLogEx(NORMAL, "  client: RRG/Iceman"); // TODO version info?
-        PrintAndLogEx(NORMAL, "\n [ PROXMARK ]");
-        PrintAndLogEx(NORMAL, "  external flash:          %s", IfPm3Flash() ? _GREEN_("present") : _YELLOW_("absent"));
-        PrintAndLogEx(NORMAL, "  smartcard reader:        %s", IfPm3Smartcard() ? _GREEN_("present") : _YELLOW_("absent"));
-        PrintAndLogEx(NORMAL, "  USART for addon support: %s\n", IfPm3FpcHost() ? _GREEN_("present") : _YELLOW_("absent"));
-
+        PrintAndLogEx(NORMAL, "\n [ PROXMARK RDV4 ]");
+        PrintAndLogEx(NORMAL, "  external flash:                  %s", IfPm3Flash() ? _GREEN_("present") : _YELLOW_("absent"));
+        PrintAndLogEx(NORMAL, "  smartcard reader:                %s", IfPm3Smartcard() ? _GREEN_("present") : _YELLOW_("absent"));
+        PrintAndLogEx(NORMAL, "\n [ PROXMARK RDV4 Extras ]");
+        PrintAndLogEx(NORMAL, "  FPC USART for BT add-on support: %s", IfPm3FpcUsartHost() ? _GREEN_("present") : _YELLOW_("absent"));
+        if (IfPm3FpcUsartDevFromUsb())
+            PrintAndLogEx(NORMAL, "  FPC USART for developer support: %s", _GREEN_("present"));
+        PrintAndLogEx(NORMAL, "");
         PrintAndLogEx(NORMAL, (char *)resp.data.asBytes);
         lookupChipID(resp.oldarg[0], resp.oldarg[1]);
     }